+++ src/main/java/com/sviryd/mikhail/repos/CityRepo.java
@@ -1,8 +1,10 @@
 package com.sviryd.mikhail.repos;
 
-import com.sviryd.mikhail.dao.entity.City;
+import com.sviryd.mikhail.entity.City;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CityRepo extends CrudRepository<City,Long>{
-    City findByName(String name);
+import java.util.Optional;
+
+public interface CityRepo extends CrudRepository<City, Long> {
+    Optional<City> findByName(String name);
 }
